add allFeatures method to mapview
