+++ lib/mapview/_mapview.mjs
@@ -12,6 +12,8 @@ import locate from './locate.mjs'
 
 import popup from './popup.mjs'
 
+import allFeatures from './allFeatures.mjs'
+
 import _highlight from './interactions/highlight.mjs'
 
 import _draw from './interactions/draw.mjs'
@@ -39,6 +41,7 @@ export default (mapview) => {
       length: (geometry) => ol.sphere.getLength(geometry),
     },
     popup,
+    allFeatures,
     layers: {},
     locations: {},
     interactions: {
