+++ lib/mapview/interactions/highlight.mjs
@@ -27,7 +27,7 @@ export default function(params){
 
     locations: new Set(),
 
-    longClickMethod: selectAll,
+    longClickMethod: mapview.allFeatures,
 
     longClickMS: 500,
 
@@ -224,7 +224,7 @@ export default function(params){
 
     if (mapview.interaction.longClick && typeof mapview.interaction.longClickMethod === 'function') {
 
-      mapview.interaction.longClickMethod(e)
+      mapview.interaction.longClickMethod(e, mapview)
       return;
     }
 
@@ -256,62 +256,6 @@ export default function(params){
     if (typeof mapview.interaction.noLocationClick === 'function') {
       mapview.interaction.noLocationClick(e)
     }
-  }
-
-  function selectAll(e) {
-
-    // Remove popup from mapview.
-    mapview.popup(null)
-
-    //const locations = [];
-
-    const features = [];
-
-    // Find locations at pixel.
-    mapview.Map.forEachFeatureAtPixel(e.pixel,
-      (F, L) => {
-
-        // const properties = F.getProperties()
-
-        // console.log(properties)
-
-        console.log(L.get('key'))
-
-        console.log(F.get('id') || F.getId())
-
-        // Push location key into locations array.
-        //locations.push(`${L.get('key')}!${F.get('id') || F.getId() || ol.util.getUid(F)}`)
-
-        features.push({F,L})
-      },
-      {
-        layerFilter: mapview.interaction.layerFilter,
-        hitTolerance: mapview.interaction.hitTolerance,
-      })
-
-    // No features at pixel.
-    if (!features.length) return;
-
-    const content = mapp.utils.html.node`
-      <div style="max-width: 66vw; max-height: 300px; overflow-x: hidden;">
-        <ul>
-        ${features.map(FL => mapp.utils.html.node`
-          <li onclick=${e => {           
-            // mapp.location.get({
-            //   layer: mapview.layers[key.split('!')[0]],
-            //   id: key.split('!')[1]
-            // })
-
-            getFeature(FL.F)
-
-            mapview.popup(null)
-          }}
-        }}>${FL.L.get('key')} []`)}`;      
-
-    mapview.popup({
-      content,
-      autoPan: true,
-    });
   }
 
   function clear() {
@@ -406,7 +350,7 @@ export default function(params){
     mapp.location.get({
       layer: feature.layer,
       table: feature.layer.table || feature.layer.tableCurrent(),
-      id: feature.layer.highlight
+      id: feature.id
     })
   }
